Added mvp for rpc server on ev3dev
This commit adds and mvp for an rpc server that can run and stop motors
on the ports on the ev3. It also includes a rpc client written in
go which runs both motors as a test for 2 seconds and then stops them.
The client/server is very minimal and have no error handling.